wifi: Handle error cases when EAP-SIM info is NULL/Unsupported
To get the SIM information,UI queries framework through the
API exposed by the Wi-Fi framework which further interacts with
the supplicant.If the supplicant times out or does not support
the EAP-SIM functionality ,it either returns null or the string
"unsupported".This throws an error while parsing at WifiEapSimInfo.
This commit handles the error cases by catching the exception.

Change-Id: Ie7b20dc32016d0396e68971415a835c40cf37989
CRs-Fixed: 796265